chore: merge components folders
