+++ rocktags/src/app/main/map/page.tsx
@@ -74,7 +74,7 @@ const buildingIcon = {
 const libraries = ["places"];
 
 import React from "react";
-import ProfileCard from "@/app/components/ProfileCard";
+import ProfileCard from "@/components/ProfileCard";
 
 
 export default function Home(): React.ReactElement {
