+++ rocktags/src/components/ui/sign-out-button.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import { signOut } from "firebase/auth";
+import { auth } from "@/config/firebase";
+import { Button } from "@/components/ui/button";
+import { useRouter } from "next/navigation";
+
+export function SignOutButton() {
+  const router = useRouter();
+
+  const handleSignOut = async () => {
+    try {
+      await signOut(auth);
+      // Clear the auth token cookie
+      document.cookie = 'firebase-auth-token=; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT';
+      router.push('/login');
+    } catch (error) {
+      console.error('Error signing out:', error);
+    }
+  };
+
+  return (
+    <Button
+      onClick={handleSignOut}
+      variant="outline"
+      size="sm"
+      className="text-sm"
+    >
+      Sign Out
+    </Button>
+  );
+}
